Fix role (#37)
* login virker nu med id som primiry key
role tabalerne virker korret med forentkey og id'er

roles mangler at blive fikset

* login virker nu med id som primiry key
role tabalerne virker korret med forentkey og id'er

roles mangler at blive fikset

* Update Routes.java

* login virker nu med id som primiry key
role tabalerne virker korret med forentkey og id'er

roles mangler at blive fikset

* tror alt virker nu

* tror alt virker nu

* det hele virker umbelbart nu

* Update Routes.java

* Update Routes.java

* userupdate virker ikek og jeg har ændert i event updater den virker nu

* userupdate virker ikek og jeg har ændert i event updater den virker nu

* Update Routes.java

---------

+++ src/main/java/org/main/routes/Routes.java
@@ -29,11 +29,12 @@ public class Routes {
         EventHandler eventHandler = new EventHandler(eventDAO);
         return () -> {
             path("events", () -> {
-                get("/", eventHandler.getAll());
-                get("/{id}", eventHandler.getById());
-                post("/", eventHandler.create(), Role.ANYONE);
-                put("/{id}", eventHandler.update());
-                delete("/{id}", eventHandler.delete());
+                before(securityHandler.authenticate());
+                get("/", eventHandler.getAll(),Role.USER,Role.ADMIN);
+                get("/{id}", eventHandler.getById(),Role.USER,Role.ADMIN);
+                post("/", eventHandler.create(), Role.ADMIN);
+                put("/{id}", eventHandler.update(), Role.ADMIN);
+                delete("/{id}", eventHandler.delete(), Role.ADMIN);
 
             });
         };
@@ -45,16 +46,17 @@ public class Routes {
         UserHandler userHandler = new UserHandler(userDAO);
         return () -> {
             path("users", () -> {
-                get(userHandler.getAllUsers());
+                before(securityHandler.authenticate());
+                get(userHandler.getAllUsers(),Role.ADMIN);
 
-                post("/user",userHandler.create());
+                post("/user",userHandler.create(),Role.ANYONE);
 
                 path("/user/{email}", () -> {
-                    get(userHandler.getByEmail());
+                    get(userHandler.getByEmail(),Role.USER,Role.ADMIN);
 
-                    put(userHandler.update());
+                    put(userHandler.update(),Role.USER);
 
-                    delete(userHandler.delete());
+                    delete(userHandler.delete(),Role.USER,Role.ADMIN);
                 });
             });
         };
@@ -69,12 +71,14 @@ public class Routes {
         return () -> {
             path("registrations", () -> {
                 before(securityHandler.authenticate());
-                get(RegistrationHandler.readAll(registrationDAO), Role.USER);
 
-                get("/id/{id}",RegistrationHandler.getRegistrationsByEventId(registrationDAO), Role.USER);
+                get(RegistrationHandler.readAll(registrationDAO), Role.ANYONE);
 
-                post("/{id}", RegistrationHandler.registerUserForEvent(registrationDAO, eventDAO, userDAO), Role.USER); // Update this line
-                delete("/{id}", RegistrationHandler.cancelUserRegistration(registrationDAO), Role.USER);
+                get("/id/{id}",RegistrationHandler.getRegistrationsByEventId(registrationDAO), Role.ANYONE);
+
+                post("/{id}", RegistrationHandler.registerUserForEvent(registrationDAO, eventDAO, userDAO), Role.ADMIN,Role.USER); // Update this line
+                delete("/{id}", RegistrationHandler.cancelUserRegistration(registrationDAO), Role.USER,Role.ADMIN);
+
 
             });
         };
@@ -86,12 +90,12 @@ public class Routes {
                 post("/login", securityHandler.login(),Role.ANYONE);
                 post("/register", securityHandler.register(),Role.ANYONE);
                 before(securityHandler.authenticate());
-                post("/reset-password", securityHandler.resetPassword(), Role.USER, Role.ADMIN);
+                post("/reset-password", securityHandler.resetPassword(), Role.ANYONE);
                 post("/logout", securityHandler.logout(), Role.USER, Role.ADMIN);
             });
         };
     }
 
-    public enum Role implements RouteRole { ANYONE, USER, ADMIN }
+    public enum Role implements RouteRole { ANYONE, USER, ADMIN}
 
 }
